requirements.txt
boto3==1.9.152
kafka-python==1.4.6
pymongo==3.8.0
selenium-wire==1.0.5
selenium==3.141.0
